refactor hdf5 reader behind open_virtual_dataset function
