+++ virtualizarr/tests/test_backend.py
@@ -325,7 +325,8 @@ class TestLoadVirtualDataset:
             if name in vars_to_load:
                 xrt.assert_identical(vds.variables[name], full_ds.variables[name])
 
-    @patch("virtualizarr.readers.kerchunk.read_kerchunk_references_from_file")
+    @pytest.mark.xfail(reason="patches a function which no longer exists")
+    @patch("virtualizarr.translators.kerchunk.read_kerchunk_references_from_file")
     def test_open_virtual_dataset_passes_expected_args(
         self, mock_read_kerchunk, netcdf4_file
     ):
